Update export declarations to meet the latest estree spec
harmony.js's dropped tests are already covered by the compare-esprima2 tests.
Close #278
