+++ test/compare-esprima2.js
@@ -25,7 +25,7 @@
 'use strict';
 
 var fs = require('fs'),
-   esprima = require('./3rdparty/esprima-2.5.0'),
+   esprima = require('./3rdparty/esprima-2.7.1'),
    escodegen = require('./loader'),
    chai = require('chai'),
    expect = chai.expect,
