+++ test/compare-esprima2/imports.expected.js
