--- test/compare-harmony/export-default-declaration.expected.js
@@ -1,2 +1,0 @@
-export default function a() {
-}
