--- test/compare-harmony/export-default-declaration.js
@@ -1,3 +1,0 @@
-export default function a () { }
-// export default var i = 20;
-// export default const K = 20;
