+++ test/harmony-esprima2.js
@@ -31,7 +31,7 @@
 
 'use strict';
 
-var esprima = require('./3rdparty/esprima-2.5.0'),
+var esprima = require('./3rdparty/esprima-2.7.1'),
     escodegen = require('./loader'),
     chai = require('chai'),
     expect = chai.expect,
