+++ test/harmony.js
@@ -2625,223 +2625,6 @@ data = {
         }
     },
 
-    'Harmony export declaration': {
-        'export function a() { }': {
-            type: 'Program',
-            body: [{
-                type: 'ExportDeclaration',
-                declaration: {
-                    type: 'FunctionDeclaration',
-                    id: {
-                        type: 'Identifier',
-                        name: 'a',
-                        range: [16, 17],
-                        loc: {
-                            start: { line: 1, column: 16 },
-                            end: { line: 1, column: 17 }
-                        }
-                    },
-                    params: [],
-                    defaults: [],
-                    body: {
-                        type: 'BlockStatement',
-                        body: [],
-                        range: [20, 23],
-                        loc: {
-                            start: { line: 1, column: 20 },
-                            end: { line: 1, column: 23 }
-                        }
-                    },
-                    rest: null,
-                    generator: false,
-                    expression: false,
-                    range: [7, 23],
-                    loc: {
-                        start: { line: 1, column: 7 },
-                        end: { line: 1, column: 23 }
-                    }
-                },
-                specifiers: null,
-                source: null,
-                range: [0, 23],
-                loc: {
-                    start: { line: 1, column: 0 },
-                    end: { line: 1, column: 23 }
-                }
-            }],
-            range: [0, 23],
-            loc: {
-                start: { line: 1, column: 0 },
-                end: { line: 1, column: 23 }
-            }
-        },
-
-        'export var i = 20': {
-            type: 'Program',
-            body: [{
-                type: 'ExportDeclaration',
-                declaration: {
-                    type: 'VariableDeclaration',
-                    declarations: [{
-                        type: 'VariableDeclarator',
-                        id: {
-                            type: 'Identifier',
-                            name: 'i',
-                            range: [11, 12],
-                            loc: {
-                                start: { line: 1, column: 11 },
-                                end: { line: 1, column: 12 }
-                            }
-                        },
-                        init: {
-                            type: 'Literal',
-                            value: 20,
-                            raw: '20',
-                            range: [15, 17],
-                            loc: {
-                                start: { line: 1, column: 15 },
-                                end: { line: 1, column: 17 }
-                            }
-                        },
-                        range: [11, 17],
-                        loc: {
-                            start: { line: 1, column: 11 },
-                            end: { line: 1, column: 17 }
-                        }
-                    }],
-                    kind: 'var',
-                    range: [7, 17],
-                    loc: {
-                        start: { line: 1, column: 7 },
-                        end: { line: 1, column: 17 }
-                    }
-                },
-                specifiers: null,
-                source: null,
-                range: [0, 17],
-                loc: {
-                    start: { line: 1, column: 0 },
-                    end: { line: 1, column: 17 }
-                }
-            }],
-            range: [0, 17],
-            loc: {
-                start: { line: 1, column: 0 },
-                end: { line: 1, column: 17 }
-            }
-        },
-
-        'export let i = 20': {
-            type: 'Program',
-            body: [{
-                type: 'ExportDeclaration',
-                declaration: {
-                    type: 'VariableDeclaration',
-                    declarations: [{
-                        type: 'VariableDeclarator',
-                        id: {
-                            type: 'Identifier',
-                            name: 'i',
-                            range: [11, 12],
-                            loc: {
-                                start: { line: 1, column: 11 },
-                                end: { line: 1, column: 12 }
-                            }
-                        },
-                        init: {
-                            type: 'Literal',
-                            value: 20,
-                            raw: '20',
-                            range: [15, 17],
-                            loc: {
-                                start: { line: 1, column: 15 },
-                                end: { line: 1, column: 17 }
-                            }
-                        },
-                        range: [11, 17],
-                        loc: {
-                            start: { line: 1, column: 11 },
-                            end: { line: 1, column: 17 }
-                        }
-                    }],
-                    kind: 'let',
-                    range: [7, 17],
-                    loc: {
-                        start: { line: 1, column: 7 },
-                        end: { line: 1, column: 17 }
-                    }
-                },
-                specifiers: null,
-                source: null,
-                range: [0, 17],
-                loc: {
-                    start: { line: 1, column: 0 },
-                    end: { line: 1, column: 17 }
-                }
-            }],
-            range: [0, 17],
-            loc: {
-                start: { line: 1, column: 0 },
-                end: { line: 1, column: 17 }
-            }
-        },
-
-        'export const i = 20': {
-            type: 'Program',
-            body: [{
-                type: 'ExportDeclaration',
-                declaration: {
-                    type: 'VariableDeclaration',
-                    declarations: [{
-                        type: 'VariableDeclarator',
-                        id: {
-                            type: 'Identifier',
-                            name: 'i',
-                            range: [13, 14],
-                            loc: {
-                                start: { line: 1, column: 13 },
-                                end: { line: 1, column: 14 }
-                            }
-                        },
-                        init: {
-                            type: 'Literal',
-                            value: 20,
-                            raw: '20',
-                            range: [17, 19],
-                            loc: {
-                                start: { line: 1, column: 17 },
-                                end: { line: 1, column: 19 }
-                            }
-                        },
-                        range: [13, 19],
-                        loc: {
-                            start: { line: 1, column: 13 },
-                            end: { line: 1, column: 19 }
-                        }
-                    }],
-                    kind: 'const',
-                    range: [7, 19],
-                    loc: {
-                        start: { line: 1, column: 7 },
-                        end: { line: 1, column: 19 }
-                    }
-                },
-                specifiers: null,
-                source: null,
-                range: [0, 19],
-                loc: {
-                    start: { line: 1, column: 0 },
-                    end: { line: 1, column: 19 }
-                }
-            }],
-            range: [0, 19],
-            loc: {
-                start: { line: 1, column: 0 },
-                end: { line: 1, column: 19 }
-            }
-        }
-    },
-
     'Harmony classes': {
         'class hello{}': {
             options: {
